refactor: replace board files with ts
