+++ .eslintrc.json
@@ -1,5 +1,9 @@
 {
   "root": true,
+  "parser": "@typescript-eslint/parser",
+  "plugins": [
+    "@typescript-eslint"
+  ],
   "env": {
     "es2020": true,
     "jasmine": true,
@@ -7,20 +11,14 @@
     "node": true
   },
   "parserOptions": {
-    "sourceType": "module",
-    "ecmaVersion": 2021,
-    "ecmaFeatures": {
-      "impliedStrict": true
-    }
+      "sourceType": "module"
   },
   "settings": {
     "noInlineConfig": true
   },
-  "parser": "babel-eslint",
   "extends": [
     "eslint:recommended",
-    "plugin:node/recommended",
-    "airbnb-base",
+    "plugin:@typescript-eslint/eslint-recommended",
     "prettier"
   ],
   "rules": {
