+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+  "ignore": ["node_modules"],
+  "watch": ["./src/server.ts"],
+  "exec": "npm start",
+  "ext": "ts"
+}
