+++ src/app.ts
@@ -1,19 +1,19 @@
-import express from 'express';
+import express, { Express } from 'express';
 import cors from 'cors';
 import swaggerUI from 'swagger-ui-express';
 import path, { dirname } from 'path';
 import { fileURLToPath } from 'url';
 import YAML from 'yamljs';
 
-import { createUserRouter, UserController, UserService, UserRepository } from './resources/users/index.js';
-import { createBoardRouter, BoardController, BoardService, BoardRepository } from './resources/boards/index.js';
-import { createTaskRouter, TaskController, TaskService, TaskRepository } from './resources/tasks/index.js';
-import { errorHandler } from './middlewares/error-handler.js';
+import { createUserRouter, UserController, UserService, UserRepository } from './resources/users';
+import { createBoardRouter, BoardController, BoardService, BoardRepository } from './resources/boards';
+import { createTaskRouter, TaskController, TaskService, TaskRepository } from './resources/tasks';
+import { errorHandler } from './middlewares';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-export const createApp = () => {
+export const createApp = (): Express => {
   const app = express();
   const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
