+++ src/common/config.ts
@@ -9,8 +9,7 @@ dotenv.config({
   path: path.join(__dirname, '../../.env')
 });
 
-export const { PORT } = process.env;
-export const { NODE_ENV } = process.env;
-export const { MONGO_CONNECTION_STRING } = process.env;
-export const { JWT_SECRET_KEY } = process.env;
-export const AUTH_MODE = process.env.AUTH_MODE === 'true';
+const { PORT, NODE_ENV, MONGO_CONNECTION_STRING, JWT_SECRET_KEY } = process.env;
+const AUTH_MODE = process.env['AUTH_MODE'] === 'true';
+
+export { PORT, NODE_ENV, MONGO_CONNECTION_STRING, JWT_SECRET_KEY, AUTH_MODE };
