--- src/database/in-memory.database.js
@@ -1,158 +1,0 @@
-export const inMemoryDatabase = {
-  users: [
-    {
-      id: 'ee2c67eb-0f65-4c9f-bb19-c2167b2ff273',
-      name: 'Olga',
-      login: 'Anahi34',
-      password: 'cb_LknYKqtv77oO',
-    },
-    {
-      id: 'cad46f08-ff8b-4496-aa9d-d375c0fd6b03',
-      name: 'Polina',
-      login: 'Aurelia_Ebert98',
-      password: 'xeER8zESj3uZxFI',
-    },
-    {
-      id: '5d71ccd7-1c41-45c3-b918-13e172180a9b',
-      name: 'Marina',
-      login: 'Bernadette.Quitzon7',
-      password: '_nkzFAj823RLisq',
-    },
-    {
-      id: '9c568b74-b8c9-456c-a5b0-794270877639',
-      name: 'Vitaly',
-      login: 'Chance.Hartmann78',
-      password: '9ZIe5dFgNYdJstA',
-    },
-    {
-      id: '5326be4f-5201-4356-ae1e-fc9dc158c49b',
-      name: 'Angelica',
-      login: 'Christina7',
-      password: 'taRBzN7Y8e447hS',
-    },
-    {
-      id: 'a08d7a8f-30ab-4703-a6ae-2f83110e2c27',
-      name: 'Svetlana',
-      login: 'Daniela72',
-      password: 'LkxH8SNBm97OZkr',
-    },
-    {
-      id: '0b236ea2-c255-49d0-8a05-555964e749a1',
-      name: 'Gregory',
-      login: 'Deondre.Breitenberg',
-      password: 'fCet3lOpi9zJiM2',
-    },
-    {
-      id: 'aa5c335d-6134-4c71-b79d-3e8ec9c3f6c9',
-      name: 'Innocenty',
-      login: 'Eddie57',
-      password: '9AP9maKTHqsBBtt',
-    },
-    {
-      id: 'f6c08f08-c8e2-4497-8e4c-54df9c3b7f7b',
-      name: 'Mark',
-      login: 'Edwin39',
-      password: 'EWOtUd15As2quOe',
-    },
-    {
-      id: 'c006e2ae-0290-4197-ad83-6316a3a829f3',
-      name: 'Nickolai',
-      login: 'Elvis_Abernathy',
-      password: '7aW2SOVdBhBduTj',
-    },
-    {
-      id: '2916492d-633e-4de2-932e-cf463714bfc0',
-      name: 'Stanislav',
-      login: 'Emanuel.Jaskolski',
-      password: 'wrH53YwNacNBWy5',
-    },
-    {
-      id: 'd257ed78-efb1-4c0b-89c9-3a1af55e26ca',
-      name: 'Victoria',
-      login: 'Gabe_Zulauf',
-      password: 'afRasY246oVdNS4',
-    },
-    {
-      id: 'b1c521c7-750f-4636-9046-d52c1fad9212',
-      name: 'Julie',
-      login: 'Iva26',
-      password: 'HACqaRQ7zLRzZC9',
-    },
-    {
-      id: '8e0e3ba3-4213-4b96-a248-29564cc288e7',
-      name: 'Eugene',
-      login: 'Izaiah_Upton2',
-      password: 'ZA5EcLLBwHXzJBn',
-    },
-    {
-      id: '96cecc7c-91ad-4a57-b8e9-7b4aab42e7ae',
-      name: 'Stephan',
-      login: 'Jamie_Lemke',
-      password: 'JLb6wC6Vkj90Tie',
-    },
-    {
-      id: 'a035ae8d-e4bd-45d8-83ed-d3e4cb6bb58e',
-      name: 'Gleb',
-      login: 'Jammie.Hagenes',
-      password: 'tpEKxPBSXY8BkXX',
-    },
-    {
-      id: 'e1a12b66-f876-4970-8ef7-2865303df0e7',
-      name: 'Semen',
-      login: 'Jorge.Ortiz',
-      password: 'kWOytbWtff5BbD7',
-    },
-    {
-      id: '20d49a89-bbaf-4939-b240-eb5e3b0eedca',
-      name: 'Dmitry',
-      login: 'Kale_Oberbrunner',
-      password: 'IhBmhq0Yl0EWhR6',
-    },
-    {
-      id: 'a065cd19-442e-441b-95ae-85defc01f339',
-      name: 'Katherine',
-      login: 'Kamren.Fisher',
-      password: '6Kw8_i2du9PE2ZT',
-    },
-    {
-      id: '75847f44-a80b-4669-940b-67b8a2e872c5',
-      name: 'Veniamin',
-      login: 'Kendall74',
-      password: '1I74jcdV7s8k8Bi',
-    },
-    {
-      id: '9688b3b1-483d-4d1a-b88a-5f26347efb8f',
-      name: 'Elizabeth',
-      login: 'Kimberly65',
-      password: 'LtsyfqD1QM_wH38',
-    },
-    {
-      id: '33b3560d-be4e-4e2f-9940-a6c59837761e',
-      name: 'Violett',
-      login: 'Ludie.Dietrich41',
-      password: 'uUs4pBxJgGrxQqL',
-    },
-    {
-      id: 'fc14fecc-c9a9-485c-93f6-1c58585a8c41',
-      name: 'Christine',
-      login: 'Mckayla_Streich',
-      password: '0p4t6WRw3mRUcwt',
-    },
-    {
-      id: 'b34b2c4b-6a3c-402a-bd01-d1d0d64695f7',
-      name: 'Anton',
-      login: 'Rahul_Hills',
-      password: '5ugRWmVHAAvoXY2',
-    },
-    {
-      id: 'a891756c-dcd7-40bd-a006-78ed8d3013e3',
-      name: 'Alexandra',
-      login: 'Rebecca.Hegmann',
-      password: 'NFhrRBtn0QiPe0k',
-    }
-  ],
-  boards: [
-  ],
-  tasks: [
-  ],
-}
