+++ src/database/in-memory.database.ts
@@ -1,0 +1,8 @@
+export const inMemoryDatabase = {
+  users: [
+  ],
+  boards: [
+  ],
+  tasks: [
+  ],
+}
