--- src/resources/boards/index.js
@@ -1,5 +1,0 @@
-export { Board } from './board.model.js';
-export { BoardMemoryRepository as BoardRepository } from './board.memory.repository.js';
-export { BoardService } from './board.service.js';
-export { BoardController } from './board.controller.js';
-export { createBoardRouter } from './board.router.js';
