+++ src/resources/boards/index.ts
@@ -1,0 +1,5 @@
+export { Board } from './board.model';
+export { BoardMemoryRepository as BoardRepository } from './board.memory.repository';
+export { BoardService } from './board.service';
+export { BoardController } from './board.controller';
+export { createBoardRouter } from './board.router';
