--- src/resources/columns/index.js
@@ -1,1 +1,0 @@
-export { Column } from './column.model.js';
