+++ src/resources/columns/index.ts
@@ -1,0 +1,1 @@
+export { Column } from './column.model';
