+++ src/types/index.ts
@@ -1,0 +1,7 @@
+import { Column } from "../resources/columns";
+
+export type TBoardBody = {
+  id: string;
+  title: string;
+  columns: Column[] | null;
+}
